+++ examples/CorePlotGallery/src/plots/CandlestickPlot.m
@@ -168,8 +168,10 @@ static const NSTimeInterval oneDay = 24 * 60 * 60;
     newGraph.legend.borderLineStyle    = newGraph.plotAreaFrame.borderLineStyle;
     newGraph.legend.cornerRadius       = 5.0;
     newGraph.legend.swatchCornerRadius = 5.0;
-    newGraph.legendAnchor              = CPTRectAnchorBottom;
-    newGraph.legendDisplacement        = CGPointMake( 0.0, self.titleSize * CPTFloat(3.0) );
+    newGraph.legend.swatchLayout       = CPTLegendSwatchLayoutRight;
+
+    newGraph.legendAnchor       = CPTRectAnchorBottom;
+    newGraph.legendDisplacement = CGPointMake( 0.0, self.titleSize * CPTFloat(3.0) );
 
     // Set plot ranges
     CPTXYPlotSpace *plotSpace = (CPTXYPlotSpace *)newGraph.defaultPlotSpace;